+++ spec/spec_helper.rb
@@ -4,6 +4,9 @@ require "stringio"
 
 require_relative "../lib/lumberjack_data_dog"
 
+Lumberjack.deprecation_mode = :raise
+Lumberjack.raise_logger_errors = true
+
 RSpec.configure do |config|
   config.warnings = true
   config.disable_monkey_patching!